+++ elevenlabs_rs/CHANGELOG.md
@@ -5,11 +5,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - `ComputeRAGIndex` endpoint
-- `prompt_injectable` field to `GetKnowledgeBaseResponse`, `CreateKnowledgeBaseResponse`, and `Document`
+- `prompt_injectable` field to `GetKnowledgeBaseDocResponse`, `CreateKnowledgeBaseDocResponse`, and `Document`
+- `name` field to `CreateKnowledgeBaseDocBody`
+- 
+### Changed
+- **Breaking**: `GetKnowledgeBase` to `GetKnowledgeBaseDoc`
+- **Breaking**: `GetKnowledgeBaseResponse` to `GetKnowledgeBaseDocResponse`
+- **Breaking**: `CreateKnowledgeBase` to `CreateKnowledgeBaseDoc`
+- **Breaking**: `CreateKnowledgeBaseBody` to `CreateKnowledgeBaseDocBody`
+- **Breaking**: `CreateKnowledgeBaseResponse` to `CreateKnowledgeBaseDocResponse`
+
 ### Removed
 - **Breaking:**`GetVoiceQuery` as now deprecated
+
 ## [0.5.1] - 2025-02-28
 ### Added
 - moved `SharedVoice` back to `voice_library` module